Crate rules

Handles all the Rule related stuff

Modules

parser

A Collection of parsers for Rule related Datatypes

rule_list

Contains the Datatypes for the Left-Right inspired List of Rules

Structs

CorsOpts

The Options to configure CORS

Middleware

A Middleware modifies a Request or Response using the provided Action

MiddlewareList

Stores a list of all Middlewares.

ReadManager

A Wrapper around the RuleList-Reader that makes it easier to find a matching Rule for a given Request

Rule

A Rule represents a single Routing-Rule, this consists of a Matcher, Priority, Middlewares and a Service

Service

A Service represents a Collection of final IP-Addresses that can receive Requests

Enums

Action

An Action performs a specific Mutation on a Request or Response

ConnectError

The Error returned by the Service when it fails to establish an outgoing connection

Matcher

Used to determine if a Request matches certain criteria

RuleTLS

The TLS-Value for a single Rule

Functions

new

Creates a new Read/Write Pair for the Rule-List